most of shift editing
