+++ src/components/ShiftDetailEditor.jsx
@@ -1,29 +1,39 @@
 import Modal from "./utils/Modal";
 import PropTypes from "prop-types";
 import { dateToHourMinute } from "../utils/dateToHourMinute";
-function ScheduleAddFullEmptyForm(props) {
+function ShiftDetailEditor(props) {
+  var shift = props.shift;
+  if (!shift) return;
+  function onSubmit(event) {
+    event.preventDefault();
+    props.onSubmit(shift);
+  }
+  function onDelete(event) {
+    event.preventDefault();
+    props.onDelete(shift);
+  }
   return (
     <Modal
-      title="Add a new shift"
+      title="Edit shift"
       isModalOpen={props.isModalOpen}
       onClose={props.closeModal}
     >
       <div id="info" className="mb-4">
-        <p className="mb-2">Event ID: {props.selectedEvent?.id}</p>
-        {props.selectedEvent?.allDay ? (
-          <p className="mb-2">All day: {props.selectedEvent?.allDay}</p>
+        <p className="mb-2">Event ID: {shift.id}</p>
+        {props.shift?.allDay ? (
+          <p className="mb-2">All day: {shift.allDay}</p>
         ) : (
           <>
             <p className="mb-2">
-              Start Time: {dateToHourMinute(props.selectedEvent?.startTime)}
+              Start Time: {dateToHourMinute(shift.startTime)}
             </p>
             <p className="mb-2">
-              End Time: {dateToHourMinute(props.selectedEvent?.endTime)}
+              End Time: {dateToHourMinute(shift.endTime)}
             </p>
           </>
         )}
       </div>
-      <form className="max-w-sm mx-auto" onSubmit={props.onSubmit}>
+      <form className="max-w-sm mx-auto" onSubmit={onSubmit}>
         <div className="mb-4">
           <label
             className="block text-gray-700 text-sm font-bold mb-2"
@@ -36,6 +46,8 @@ function ScheduleAddFullEmptyForm(props) {
             id="name"
             type="text"
             name="name"
+            defaultValue={shift.name}
+            onChange={(e) => {shift.name = e.target.value}}
           />
         </div>
         <div className="mb-4">
@@ -50,6 +62,8 @@ function ScheduleAddFullEmptyForm(props) {
             id="desc"
             type="text"
             name="desc"
+            defaultValue={shift.desc}
+            onChange={(e) => {shift.desc = e.target.value}}
           />
         </div>
         <div className="mb-4">
@@ -64,6 +78,8 @@ function ScheduleAddFullEmptyForm(props) {
             id="location"
             type="text"
             name="location"
+            defaultValue={shift.location}
+            onChange={(e) => {shift.desc = e.target.value}}
           />
         </div>
         <input
@@ -71,16 +87,24 @@ function ScheduleAddFullEmptyForm(props) {
           value="Submit"
           className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
         />
+        <button
+          onClick={onDelete}
+          className="bg-purple-500 hover:bg-purple-700 text-white font-bold py-2 px-4 rounded-full ml-4"
+        >
+          Delete shift
+        </button>
       </form>
     </Modal>
   );
 }
 
-ScheduleAddFullEmptyForm.propTypes = {
+ShiftDetailEditor.propTypes = {
   isModalOpen: PropTypes.bool,
   closeModal: PropTypes.func,
   onSubmit: PropTypes.func,
-  selectedEvent: PropTypes.object,
+  onDelete: PropTypes.func,
+  shift: PropTypes.object,
+  key: PropTypes.number,
 };
 
-export default ScheduleAddFullEmptyForm;
+export default ShiftDetailEditor;
