+++ src/routes/ScheduleEditor.tsx
@@ -4,13 +4,12 @@ import Schedule from "../components/Schedule";
 import { getUser } from "../api/getUser";
 import { useAtom } from "jotai";
 import { userIdAtom } from "../globalAtom";
-import TestSchedule from "../components/TestSchedule";
 
 const ScheduleView = () => {
   let { employee } = useParams();
   return (
-    <div className="min-h-screen">
-      <TestSchedule employeeId={employee ?? ""} />
+    <div className="">
+      <Schedule employeeId={employee ?? ""} />
     </div>
   );
 };
